refactor: simplify register.js (add codeblocks component and services)
